File/Folder structure modified
1. file/folder structure modified
2. utf-8 added
3. simple note added
